+++ packages/platform-android/src/__tests__/touch-helper.fixtures.ts
@@ -2,7 +2,7 @@ import { ANDROID_EMULATOR } from './test-utils/device-fixtures.ts';
 import { buildGesturePlan } from '@agent-device/contracts/gesture-plan';
 import type { DeviceInfo } from '@agent-device/kernel/device';
 import type { AndroidAdbExecutor } from '../adb-executor.ts';
-import type { AndroidLongPressTouchPlan } from '../touch-plan.ts';
+import type { AndroidLongPressTouchPlan } from '../touch-plan-lowering.ts';
 
 // The one-shot touch helper path now runs through the snapshot-helper APK/runner
 // (issue #1275 consolidation), so these fixtures mirror the snapshot helper's
