--- packages/platform-android/src/touch-plan.ts
@@ -1,6 +1,0 @@
-export {
-  lowerAndroidTouchPlan,
-  type AndroidLongPressTouchPlan,
-  type AndroidLoweredTouchPlan,
-  type AndroidTouchPlan,
-} from '@agent-device/contracts/android-touch-plan';
